Model generator which takes input data and populates the transition matrix.
Main part is one function to take a list of states and a placeholder for the matrix of transition probabilities and to populate the matrix and then normalize each row to sum to unity.
There's also some code to read data from a file and then call the function on that input data.
Some code used for testing is also pasted along at the bottom, and commented out.